requirements.txt
Flask==3.0.2
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
SQLAlchemy==2.0.38
Werkzeug==3.0.1
marshmallow==3.21.1
python-json-logger==2.0.7
requests==2.31.0 